pneumonia_detection - Copy/backend/requirements.txt
# ===== BACKEND =====
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Mail==0.9.1

# ===== ML CORE =====
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.2.2
matplotlib==3.7.1
Pillow==10.0.0

# ===== DEVELOPMENT =====
jupyterlab==4.0.2  # Optional
kagglehub==1.0.0   # Optional for dataset downloads